requirements.txt
certifi==2025.10.5
charset-normalizer==3.4.4
idna==3.11
kafka-python==2.0.2
numpy==1.26.4
pandas==2.1.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
pytz==2025.2
requests==2.31.0
six==1.17.0
tzdata==2025.2
urllib3==2.5.0
certifi==2025.10.5
charset-normalizer==3.4.4
idna==3.11
kafka-python==2.0.2
numpy==1.26.4
pandas==2.1.4
psycopg2-binary==2.9.11
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
pytz==2025.2
requests==2.31.0
six==1.17.0
tzdata==2025.2
urllib3==2.5.0
psycopg2-binary==2.9.11